+++ CHANGELOG.md
@@ -6,6 +6,16 @@ until we achieve a stable v1.0 release
 
 ## v0.1.2 - unreleased
 
+- 💥 BREAKING: Removed the shadow DOM content wrapper,
+  and all shadow DOM styles
+- 💥 BREAKING: Renamed and added control-panel parts,
+  to allow for more customization of the default panel
+- 🚀 NEW: Default styles are in `slide-deck.css`
+  and can be applied from the light DOM
+- 🚀 NEW: The entire control panel can be replaced
+  from the light DOM using `slot=control-panel`
+  on a slotted `dialog` element
+- 🐞 FIXED: Slotted controls are no longer treated as slides
 - 🚀 NEW: When `key-control` is activated (including on-load),
   we target the stored active slide (or the first slide)
 - 🐞 FIXED: When restoring the active slide from memory,